Renovate the config files
