--- .eslintrc.json
@@ -1,27 +1,0 @@
-{
-  "env": {
-    "browser": true,
-    "es2021": true,
-    "node": true
-  },
-  "extends": [
-    "eslint:recommended",
-    "plugin:vue/vue3-recommended",
-    "plugin:@typescript-eslint/recommended",
-    "@nuxtjs/eslint-config-typescript",
-    "prettier"
-  ],
-  "parserOptions": {
-    "ecmaVersion": "latest",
-    "parser": "@typescript-eslint/parser",
-    "sourceType": "module"
-  },
-  "plugins": ["vue", "@typescript-eslint"],
-  "rules": {
-    "no-console": "off",
-    "vue/no-v-html": "off",
-    "comma-dangle": ["error", "only-multiline"],
-    "vue/multi-word-component-names": "off",
-    "import/no-named-as-default": "off"
-  }
-}
