+++ .prettierrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  printWidth: 100,
+  semi: false,
+  singleQuote: true,
+  endOfLine: 'auto',
+}
