--- .prettierrc.json
@@ -1,6 +1,0 @@
-{
-  "printWidth": 100,
-  "semi": false,
-  "singleQuote": true,
-  "endOfLine": "auto"
-}
