+++ package.json
@@ -6,23 +6,22 @@
     "generate": "nuxi generate",
     "preview": "nuxi preview",
     "postinstall": "nuxi prepare",
-    "lint:js": "eslint --ext .js,.vue,.ts,.jsx,.tsx,.html,.vue .",
+    "lint:js": "eslint . --ext .js,.vue,.ts",
     "lint": "npm run lint:js && prettier --check .",
-    "fix": "eslint --fix --ext .js,.vue,.ts,.jsx,.tsx,.html,.vue ."
+    "fix": "eslint . --fix --ext .js,.vue,.ts"
   },
   "devDependencies": {
     "@nuxtjs/eslint-config-typescript": "^12.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.50.0",
-    "@typescript-eslint/parser": "^5.50.0",
+    "@types/node": "^18.16.0",
     "eslint": "^8.33.0",
     "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-vue": "^9.9.0",
+    "eslint-plugin-prettier": "^4.2.1",
     "nuxt": "^3.4.1",
     "prettier": "^2.8.3",
     "sass": "^1.58.0",
-    "typescript": "^4.9.5",
-    "vite-plugin-eslint": "^1.8.1",
-    "vite-svg-loader": "^4.0.0"
+    "typescript": "^5.0.4",
+    "vite-svg-loader": "^4.0.0",
+    "vue-tsc": "^1.4.4"
   },
   "dependencies": {
     "@ltd/j-toml": "^1.38.0",
