+++ pages/[type]/[id]/settings/links.vue
@@ -120,7 +120,7 @@
 </template>
 
 <script>
-import Multiselect from 'vue-multiselect'
+import { Multiselect } from 'vue-multiselect'
 import SaveIcon from '~/assets/images/utils/save.svg'
 
 export default defineNuxtComponent({
