+++ pages/settings/index.vue
@@ -135,7 +135,7 @@
 </template>
 
 <script>
-import Multiselect from 'vue-multiselect'
+import { Multiselect } from 'vue-multiselect'
 
 export default defineNuxtComponent({
   components: {
